optimize package name
