+++ agent/log/log-to-slf4j/src/main/java/org/apache/shardingsphere/agent/log/slf4j/SLF4JAgentLoggerFactory.java
@@ -17,8 +17,8 @@
 
 package org.apache.shardingsphere.agent.log.slf4j;
 
-import org.apache.shardingsphere.agent.log.AgentLogger;
-import org.apache.shardingsphere.agent.log.IAgentLoggerFactory;
+import org.apache.shardingsphere.agent.log.api.AgentLogger;
+import org.apache.shardingsphere.agent.log.spi.IAgentLoggerFactory;
 import org.slf4j.LoggerFactory;
 
 import java.util.Map;
